+++ src/main/java/com/codegym/model/ProductForm.java
@@ -3,7 +3,7 @@ package com.codegym.model;
 import org.springframework.web.multipart.MultipartFile;
 
 public class ProductForm {
-    private int id;
+    private Long id;
 
     private String name;
 
@@ -16,7 +16,7 @@ public class ProductForm {
     public ProductForm() {
     }
 
-    public ProductForm(int id, String name, double price, String description, MultipartFile image) {
+    public ProductForm(Long id, String name, double price, String description, MultipartFile image) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -24,11 +24,11 @@ public class ProductForm {
         this.image = image;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
